Update BulkPrint.ps1
- try using tasks to run windows printing concurrently
- add mac printer install
- add mac printer bulk printing

+++ BulkPrint.ps1
@@ -1,27 +1,54 @@
 # This requires PDFtoPrinter.exe
 # RB 
-# $Copiers = `
-# @{
-#     'RB Media Tent'    = '10.1.1.223';
-#     'RB South Press 1' = '10.1.1.215';
-#     'RB South Press 2' = '10.1.1.216';
-#     'RB North Press 1' = '10.1.1.217';
-#     'RB North Press 2' = '10.1.1.218';
-# }
-# CFP Indianapolis
-$Copiers = `
-@{
-	'CFP Press 1'   = '172.26.4.11'; # Press Box
-	'CFP Press 2'   = '172.26.4.12';
-	'CFP Digital 1' = '172.26.4.15'; # Event Level
-	'CFP Digital 2' = '172.26.4.16';
-	'CFP Aux 1'     = '172.26.4.13'; # Terrace Level
-	'CFP Aux 2'     = '172.26.4.14';
+ $Copiers = `
+ @{
+     'RB Media Tent'    = '10.1.1.223';
+     'RB South Press 1' = '10.1.1.215';
+     'RB South Press 2' = '10.1.1.216';
+     'RB North Press 1' = '10.1.1.217';
+     'RB North Press 2' = '10.1.1.218';
+     'RB Digital Media' = '10.1.128.229';
+ }
+ $MacDriverPath = '/Library/Printers/PPDs/Contents/Resources/'
+ $MacCopiers = @(
+            @{ Name = "RB_Media_Tent";    "IP" = "10.1.1.223";   "Driver" = "/Library/Printers/PPDs/Contents/Resources/SHARP MX-M623N.PPD.gz" },
+            @{ Name = "RB_South_Press_1"; "IP" = "10.1.1.215";   "Driver" = "/Library/Printers/PPDs/Contents/Resources/SHARP MX-M753N.PPD.gz" },
+            @{ Name = "RB_South_Press_2"; "IP" = "10.1.1.216";   "Driver" = "/Library/Printers/PPDs/Contents/Resources/SHARP MX-M753N.PPD.gz" },
+            @{ Name = "RB_North_Press_1"; "IP" = "10.1.1.217";   "Driver" = "/Library/Printers/PPDs/Contents/Resources/SHARP MX-M753N.PPD.gz" },
+            @{ Name = "RB_North_Press_2"; "IP" = "10.1.1.218";   "Driver" = "/Library/Printers/PPDs/Contents/Resources/SHARP MX-M753N.PPD.gz" },
+            @{ Name = "RB_Digital_Media"; "IP" = "10.1.128.229"; "Driver" = "/Library/Printers/PPDs/Contents/Resources/SHARP MX-M623N.PPD.gz" }
+ )
+foreach($Copier in $MacCopiers) {
+    Set-Variable -Name $Copier.Name -Value $Copier
 }
+# CFP Indianapolis
+#$Copiers = `
+#@{
+#	'CFP Press 1'   = '172.26.4.11'; # Press Box
+#	'CFP Press 2'   = '172.26.4.12';
+#	'CFP Digital 1' = '172.26.4.15'; # Event Level
+#	'CFP Digital 2' = '172.26.4.16';
+#	'CFP Aux 1'     = '172.26.4.13'; # Terrace Level
+#	'CFP Aux 2'     = '172.26.4.14';
+#}
 
 $DriverPath = 'C:\Users\wesgr\Downloads\WHQL Universal Print Driver_64bit'
 $DriverFile = 'sfweMENU.inf'
 $DriverName = 'SHARP UD2 PCL6'
+
+Function Install-MacPrinter($Name, $IP, $Driver) {
+    lpadmin -p $Name -L '' -E -v lpd://$IP -P $Driver
+}
+
+Function Invoke-MacPrint($Name, $File) {
+    lp -d $Name $File
+}
+
+Function Invoke-MacBroadcastPrint($Copiers=$MacCopiers, $file) {
+    foreach($Copier in $Copiers) {
+        Invoke-MacPrint -Name $Copier.Name -File $file
+    }
+}
 
 Function Add-BulkPrinterDriver($DriverPath, $DriverFile, $DriverName)
 {
@@ -89,12 +116,16 @@ Function Invoke-PrintQueue
         $path,
         $printers = $Copiers.Keys
     )
-    $files = Get-ChildItem $path | Select-Object -ExpandProperty FullName
-    foreach($file in $files)
-    {
-        Invoke-BroadcastPrint -printers $printers -file $files
+    $files = Get-ChildItem $path | Select-Object -ExpandProperty FullName | Where-Object {$_ -like '*.pdf'}
+    if (-not $files) {
+        Write-Warning "No Files in $path"
+        return
     }
 
+    foreach($file in $files)
+    {
+         Invoke-BroadcastPrint -printers $printers -file $file
+    }
     Remove-Item $files
 }
 
@@ -123,11 +154,14 @@ Function Invoke-BroadcastPrint
 
     if($printers -and $file)
     {
+	$jobs = New-Object System.Collections.ArrayList
         foreach($printer in $printers)
         {
             "`"$Path\$printer.exe`" `"$file`" `"$printer`""
-            Invoke-Expression "&`"$Path\$printer.exe`" `"$file`" `"$printer`""
+            $jobs.Add( $(Start-Job -ScriptBlock { Invoke-Expression "&`"$Using:Path\$Using:printer.exe`" `"$Using:file`" `"$Using:printer`""} ))
         }
+        Wait-Job $jobs | Remove-Job
+	Start-Sleep -Seconds 20
     }
 }
 $AcroRd32 = "'C:\Program Files (x86)\Adobe\Acrobat Reader DC\Reader\AcroRd32.exe'"
